chore: rename navigator package to executor
