+++ internal/executor/change_page.go
@@ -1,4 +1,4 @@
-package navigator
+package executor
 
 import (
 	"kurohelper/internal/cache"
@@ -7,18 +7,18 @@ import (
 	"github.com/bwmarrin/discordgo"
 )
 
-// 處理「回到首頁」的導覽邏輯。
+// 處理「列表翻頁」的導覽邏輯。
 //
 // 參數：
 //   - s: 用於發送回應的 Discord session
-//   - i: 觸發導覽的 interaction
-//   - backToHomeCID: 由呼叫端透過 cid.ToBackToHomeCIDV2() 取得
+//   - i: 觸發翻頁的 interaction
+//   - pageCID: 由呼叫端透過 cid.ToPageCIDV2() 取得
 //   - store: 存放列表資料的快取儲存
-//   - builder: 從快取資料建構訊息元件的函數。接收 (cacheValue, pageNumber, cacheID)，回傳列表第一頁的元件
-func BackToHome[T any](
+//   - builder: 從快取資料建構訊息元件的函數。接收 (cacheValue, pageNumber, cacheID)，回傳該頁的元件
+func ChangePage[T any](
 	s *discordgo.Session,
 	i *discordgo.InteractionCreate,
-	backToHomeCID *utils.BackToHomeCIDV2,
+	pageCID *utils.PageCIDV2,
 	store *cache.CacheStoreV2[T],
 	builder func(T, int, string) ([]discordgo.MessageComponent, error),
 ) {
@@ -27,7 +27,7 @@ func BackToHome[T any](
 		Type: discordgo.InteractionResponseDeferredMessageUpdate,
 	})
 
-	cidCacheValue, err := cache.CIDStore.Get(backToHomeCID.CacheID)
+	cidCacheValue, err := cache.CIDStore.Get(pageCID.CacheID)
 	if err != nil {
 		utils.HandleErrorV2(err, s, i, utils.InteractionRespondEditComplex)
 		return
@@ -39,10 +39,11 @@ func BackToHome[T any](
 		return
 	}
 
-	components, err := builder(cacheValue, 1, backToHomeCID.CacheID)
+	components, err := builder(cacheValue, pageCID.Value, pageCID.CacheID)
 	if err != nil {
 		utils.HandleErrorV2(err, s, i, utils.InteractionRespondEditComplex)
 		return
 	}
-	utils.InteractionRespondEditComplex(s, i, components)
+
+	utils.WebhookEditRespond(s, i, components)
 }
